Fix obscure corner case in conjecturing
The concrete example was a theory containing both EAssignment and EAdd
as two constructors that can appear in the same typing context.
Here, CSD fast forwards them to 'EA', then samples one disambiguating
character. If it happens to be 'd' (going towards EAdd), it turns out
that 'd' is also exactly the remaining suffix here. So the algorithm
was terminating early, producing 'EAd'.